+++ src/Fable.BluePrint.Core/Core/Button.fs
@@ -25,7 +25,7 @@ type IButtonProps =
     | Minimal of bool
     | OnClick of (MouseEvent -> unit)
     | RightIcon of string //IconName will come
-    | Text of string
+    | Text of ReactElement
     | Type of ButtonType
     interface IHTMLProp
 
@@ -34,7 +34,6 @@ module Button =
     let inline button (props : IHTMLProp list) (elems : ReactElement list) : ReactElement =
         ofImport "Button" "@blueprintjs/core"
             (keyValueList CaseRules.LowerFirst props) elems
-    let inline anchorButton (props : IHTMLProp list)
-               (elems : ReactElement list) : ReactElement =
+    let inline anchorButton (props : IHTMLProp list) (elems : ReactElement list) : ReactElement =
         ofImport "AnchorButton" "@blueprintjs/core"
             (keyValueList CaseRules.LowerFirst props) elems